requirements.txt
py-cord[voice]
discord-pretty-help
requests